power: Allow configuration of preferred backlight device
If a system contains more than one device of the same type to control
the backlight, then gnome would choose the first device of that type,
with no way for the user to override.  This change allows the user to
configue a preferred device by name, which can be used to break ties
between devices of the same type, or to override the default priority
based on the device type.

https://bugzilla.gnome.org/show_bug.cgi?id=640627
https://bugzilla.gnome.org/show_bug.cgi?id=688052
https://bugzilla.gnome.org/show_bug.cgi?id=739531
https://bugzilla.gnome.org/show_bug.cgi?id=773913